# Fix: Prevent zsh terminal crashes from arithmetic operations