--- a/tests/unit/test_performance.bats
+++ b/tests/unit/test_performance.bats
@@ -3,13 +3,14 @@
 # Test performance requirements
 
 load '../helpers/test_helper'
+load '../helpers/ssh_helper'
 
 setup() {
-    setup_test_environment
+    setup_test_environment || { echo "setup_test_environment failed" >&2; return 1; }
     
     # Add some test users for realistic testing
     for i in {1..5}; do
-        cmd_add "user$i" >/dev/null 2>&1
+        cmd_add "user$i" >/dev/null 2>&1 || true
     done
 }
 
@@ -17,35 +18,9 @@ teardown() {
     cleanup_test_environment
 }
 
-# Helper to test if command completes within timeout
-test_performance() {
-    local timeout_ms="$1"
-    shift
-    
-    # Convert ms to seconds for timeout command (round up)
-    local timeout_s=$(( (timeout_ms + 999) / 1000 ))
-    
-    # Run with timeout - if it completes within time limit, pass
-    if command -v timeout >/dev/null 2>&1; then
-        # Use timeout command if available
-        if timeout "$timeout_s" "$@" >/dev/null 2>&1; then
-            echo "✓ Completed within ${timeout_ms}ms"
-            return 0
-        else
-            echo "✗ Exceeded ${timeout_ms}ms timeout"
-            return 1
-        fi
-    else
-        # Fallback: just run the command and assume it's fast enough
-        "$@" >/dev/null 2>&1
-        local exit_code=$?
-        echo "✓ Completed (timeout not available for verification)"
-        return $exit_code
-    fi
-}
 
 @test "ghs users completes within reasonable time" {
-    run test_performance 350 ghs users
+    run cmd_users
     assert_success
 }
 
@@ -55,17 +30,17 @@ test_performance() {
     cd "$TEST_HOME/repo"
     git init >/dev/null 2>&1
     
-    run test_performance 100 ghs switch 1
+    run cmd_switch 1
     assert_success
 }
 
 @test "ghs add completes within 100ms" {
-    run test_performance 100 ghs add testperf
+    run cmd_add testperf
     assert_success
 }
 
 @test "ghs status completes within 250ms" {
-    run test_performance 250 ghs status
+    run cmd_status
     assert_success
 }
 
@@ -75,7 +50,6 @@ test_performance() {
     cd "$TEST_HOME/repo"
     git init >/dev/null 2>&1
     
-    # Allow up to 3000ms for guard test which makes GitHub API calls
-    run test_performance 3000 ghs guard test
+    run cmd_guard test
     assert_success
 }