--- a/tests/unit/test_performance.bats
+++ b/tests/unit/test_performance.bats
@@ -18,10 +18,29 @@ teardown() {
     cleanup_test_environment
 }
 
+# Helper to measure command execution time in milliseconds
+measure_time_ms() {
+    local start_ns=$(date +%s%N 2>/dev/null || echo "0")
+    if [[ "$start_ns" == "0" ]]; then
+        # macOS doesn't support nanoseconds, use python
+        local start_ms=$(python3 -c 'import time; print(int(time.time() * 1000))')
+        "$@" >/dev/null 2>&1
+        local end_ms=$(python3 -c 'import time; print(int(time.time() * 1000))')
+        echo $((end_ms - start_ms))
+    else
+        # Linux supports nanoseconds
+        "$@" >/dev/null 2>&1
+        local end_ns=$(date +%s%N)
+        echo $(( (end_ns - start_ns) / 1000000 ))
+    fi
+}
+
 
 @test "ghs users completes within reasonable time" {
-    run cmd_users
-    assert_success
+    local duration=$(measure_time_ms cmd_users)
+    echo "# Duration: ${duration}ms" >&3
+    # Allow up to 350ms for bash script startup overhead with profile lookups including host info
+    [[ "$duration" -lt 350 ]]
 }
 
 @test "ghs switch completes within 100ms" {
@@ -30,18 +49,21 @@ teardown() {
     cd "$TEST_HOME/repo"
     git init >/dev/null 2>&1
     
-    run cmd_switch 1
-    assert_success
+    local duration=$(measure_time_ms cmd_switch 1)
+    echo "# Duration: ${duration}ms" >&3
+    [[ "$duration" -lt 100 ]]
 }
 
 @test "ghs add completes within 100ms" {
-    run cmd_add testperf
-    assert_success
+    local duration=$(measure_time_ms cmd_add testperf)
+    echo "# Duration: ${duration}ms" >&3
+    [[ "$duration" -lt 100 ]]
 }
 
 @test "ghs status completes within 250ms" {
-    run cmd_status
-    assert_success
+    local duration=$(measure_time_ms cmd_status)
+    echo "# Duration: ${duration}ms" >&3
+    [[ "$duration" -lt 250 ]]
 }
 
 @test "ghs guard test completes within reasonable time" {
@@ -50,6 +72,8 @@ teardown() {
     cd "$TEST_HOME/repo"
     git init >/dev/null 2>&1
     
-    run cmd_guard test
-    assert_success
+    local duration=$(measure_time_ms cmd_guard test)
+    echo "# Duration: ${duration}ms" >&3
+    # Allow up to 3000ms for guard test which makes GitHub API calls
+    [[ "$duration" -lt 3000 ]]
 }