… .github/workflows/ci.yml
… CLAUDE.md
… Documentation/INVESTIGATE-TEST-NUMBERING-ISSUE.md
… Documentation/Plans/E2E-TERMINAL-TESTING-PLAN.md
… Documentation/Plans/FIX-PLAN-ZshPathCompatibility.md
… Documentation/Plans/IMPLEMENTATION-PLAN-DirectoryAutoSwitch.md
… tests/unit/test_zsh_compatibility.bats