## Summary
- Fixed zsh terminal crashes when running `ghs` command
- Fixed CI test execution failures (5 performance tests not running)
- Restored proper performance measurements in tests

## Problem
1. **Terminal crash**: When running `ghs` in zsh, the terminal would crash with exit code 127 due to arithmetic operations with `set -e`
2. **CI failure**: Performance tests weren't executing, causing "Executed 0 instead of expected 5 tests" error

## Root Cause
1. **Zsh crash**: Arithmetic expressions like `((var++))` return 0 when incrementing from 0 to 1, causing zsh with `set -e` to exit
2. **Test failure**: Performance tests were missing required `load '../helpers/ssh_helper'` statement

## Solution
1. **Arithmetic fix**: Changed all `((var++))` to `var=$((var + 1))` throughout codebase (8 instances)
2. **Test fix**: 
   - Added missing `ssh_helper` load to performance tests
   - Restored `measure_time_ms` function and actual performance measurements
   - Tests now properly verify execution times (100ms-3000ms thresholds)

## Testing
- ✅ Manual testing: `ghs` command works in zsh without crashes
- ✅ All 174 tests pass with proper performance measurements
- ✅ CI checks pass completely

## Test plan
- [x] Run `ghs` in a fresh zsh shell - should not crash
- [x] Run `npm test` - all 174 tests should execute and pass
- [x] Check performance test output shows actual durations
- [x] Run `npm run ci-check` - should pass all checks

🤖 Generated with [Claude Code](https://claude.ai/code)