@seconds-0 2025-07-14
## CI Update

Added performance test threshold adjustments for CI environments. Ubuntu CI runners can be slower than local environments, causing timing-sensitive tests to fail.

The fix detects when running in CI (via `CI` environment variable) and allows extra time:
- ghs users: 350ms → 500ms in CI
- ghs switch: 100ms → 200ms in CI  
- ghs add: 100ms → 200ms in CI
- ghs status: 250ms → 400ms in CI

This matches the existing pattern for guard tests (3000ms in CI) and should resolve the Ubuntu CI failures while maintaining performance standards for local development.