@seconds-0 2025-07-14
## CI Fix Applied

The CI failures were due to zsh not being installed in the Ubuntu CI environment. Our new zsh compatibility tests require zsh to be available.

**Fix**: Added `zsh` to the apt-get install list in the CI workflow.

This should resolve all CI failures. The tests were already passing locally because zsh is installed on development machines.